+++ redis/src/pipeline_wrapper.rs
@@ -1,7 +1,6 @@
 use std::ops::{Deref, DerefMut};
 
 use redis::{FromRedisValue, RedisResult, ToRedisArgs};
-use tokio_compat_02::FutureExt;
 
 use crate::{Cmd, ConnectionWrapper};
 
@@ -61,12 +60,11 @@ impl Pipeline {
     ) -> RedisResult<T> {
         self.pipeline
             .query_async(DerefMut::deref_mut(con))
-            .compat()
             .await
     }
     /// See [redis::Pipeline::execute_async](https://docs.rs/redis/latest/redis/struct.Pipeline.html#method.execute_async)
     pub async fn execute_async(&self, con: &mut ConnectionWrapper) -> RedisResult<()> {
-        self.query_async::<redis::Value>(con).compat().await?;
+        self.query_async::<redis::Value>(con).await?;
         Ok(())
     }
 }