+++ redis/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Update to `deadpool 0.6` (`tokio 0.3`)
 * Disable `redis` default features
+* Mark `Config::from_env` as deprecated
 
 ## v0.6.1